[chore] Move memorylimiter dependencies under one module
