+++ internal/memorylimiter/iruntime/total_memory_other.go
@@ -3,7 +3,7 @@
 
 //go:build !linux
 
-package iruntime // import "go.opentelemetry.io/collector/internal/iruntime"
+package iruntime // import "go.opentelemetry.io/collector/internal/memorylimiter/iruntime"
 
 // TotalMemory returns total available memory for non-linux platforms.
 func TotalMemory() (uint64, error) {
