+++ internal/memorylimiter/memorylimiter.go
@@ -15,7 +15,7 @@ import (
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/internal/iruntime"
+	"go.opentelemetry.io/collector/internal/memorylimiter/iruntime"
 )
 
 const (
